refactor(sysadmin): extract pure functions into classifiers / aggregations / pagination modules
Move 9 pure functions (classifyMember, isDormant, computeStageCounts,
computeStageBreakdown, computeTenureDistribution, computeDormantCount,
computeCohortFunnel, computeActivityRate3mAvg, paginateMembers) and
their associated types / constants out of SysAdminService class into
3 module files. Class methods now delegate to the new module functions
so callers (usecase.ts) keep working unchanged. service.ts re-exports
the moved symbols so existing import paths via the service module
continue to resolve; direct imports from the new modules and the
class-method removal land in follow-up commits.
