+++ src/application/domain/sysadmin/service.ts
@@ -8,264 +8,69 @@ import {
 } from "@/application/domain/sysadmin/data/type";
 import {
   addDays,
-  bigintToSafeNumber,
   isoWeekStartJst,
   jstMonthStart,
   jstMonthStartOffset,
   jstNextMonthStart,
   percentChange,
-  truncateToJstDate,
 } from "@/application/domain/report/util";
 import { asOfBounds } from "@/application/domain/sysadmin/bounds";
+import { SegmentThresholds } from "@/application/domain/sysadmin/classifiers";
+import {
+  computeActivityRate3mAvg as computeActivityRate3mAvgImpl,
+  computeCohortFunnel as computeCohortFunnelImpl,
+  computeDormantCount as computeDormantCountImpl,
+  computeStageBreakdown as computeStageBreakdownImpl,
+  computeStageCounts as computeStageCountsImpl,
+  computeTenureDistribution as computeTenureDistributionImpl,
+  rateOf,
+  StageBreakdown,
+  StageCounts,
+  SysAdminCohortFunnelPoint,
+  TenureDistribution,
+  WeeklyRetentionPoint,
+  MonthlyCohortPoint,
+  CHAIN_DEPTH_MAX_BUCKET,
+} from "@/application/domain/sysadmin/aggregations";
+import {
+  MemberListParams,
+  MemberListResult,
+  paginateMembers as paginateMembersImpl,
+} from "@/application/domain/sysadmin/pagination";
 
-/**
- * Stage-count thresholds come from the client. tier1 >= tier2 >= 0 is
- * enforced at the service boundary so the cumulative-count invariant
- * (`tier2Count >= tier1Count`) is guaranteed downstream.
- */
-export type SegmentThresholds = {
-  tier1: number;
-  tier2: number;
-  /**
-   * Minimum tenure required for tier1 / tier2 eligibility, expressed
-   * in calendar months for the operator-facing API but evaluated
-   * internally as `daysIn >= minMonthsIn × 30`. Filters out the
-   * short-tenure artifact where a brand new member with one donation
-   * gets `userSendRate = 1/1 = 1.0` and is auto-classified as habitual.
-   * `activeCount` and `passiveCount` are unaffected (they're tenure-
-   * independent by construction).
-   *
-   * The day-based check is intentional: monthsIn is calendar-month
-   * based and inflates a 2-day cross-month-boundary tenure to
-   * monthsIn = 2, which would silently pass `minMonthsIn = 2`. See
-   * `classifyMember` for the full reasoning.
-   */
-  minMonthsIn: number;
-};
-
-export const DEFAULT_SEGMENT_THRESHOLDS: SegmentThresholds = {
-  tier1: 0.7,
-  tier2: 0.4,
-  minMonthsIn: 1,
-};
-
-export const MIN_MIN_MONTHS_IN = 1;
-export const MAX_MIN_MONTHS_IN = 120;
-
-/**
- * Result of classifying a community's members against the supplied
- * thresholds. `tier2Count` and `tier1Count` are cumulative (tier2
- * INCLUDES tier1 members); the per-bucket breakdown lives in
- * `stageBreakdown` below so both shapes are available without a second
- * scan.
- */
-export type StageCounts = {
-  total: number;
-  tier1Count: number;
-  tier2Count: number;
-  activeCount: number;
-  passiveCount: number;
-};
-
-export type StageBucketStats = {
-  count: number;
-  pct: number;
-  pointsContributionPct: number;
-  avgSendRate: number;
-  avgMonthsIn: number;
-};
-
-export type StageBreakdown = {
-  habitual: StageBucketStats;
-  regular: StageBucketStats;
-  occasional: StageBucketStats;
-  latent: StageBucketStats;
-};
-
-/**
- * Tenure-bucket distribution of a community's members at asOf,
- * classified on `daysIn` (JST calendar-day tenure). Buckets are
- * mutually exclusive and exhaustive; the four counts always sum to
- * the input `members.length`. Boundaries are intentionally
- * day-based (not month-based) to side-step the GREATEST(1, ...)
- * floor that `monthsIn` carries.
- *
- * `monthlyHistogram` carries the L3 deep-dive breakdown into 13
- * monthly buckets (0..12, where 12 aggregates 12+ months). Same
- * member set, finer granularity. The four coarse buckets remain
- * authoritative for L1 / L2 display; the histogram is additional.
- */
-export type TenureHistogramBucket = {
-  monthsIn: number;
-  count: number;
-};
-
-export type TenureDistribution = {
-  lt1Month: number;
-  m1to3Months: number;
-  m3to12Months: number;
-  gte12Months: number;
-  monthlyHistogram: TenureHistogramBucket[];
-};
-
-/** Number of monthly histogram buckets for the L3 deep-dive
- * (`monthsIn` 0..12; 12 aggregates 12+). Pinned as a constant so
- * the service computation and the test fixtures agree on the array
- * shape. */
-export const TENURE_MONTHLY_BUCKETS = 13;
-
-/** Maximum chain-depth bucket (inclusive). The L3
- * chainDepthDistribution emits depth 1..N where N aggregates
- * `chain_depth >= N`. 5 chosen as a starting point; revisit if
- * real-data inspection of `maxChainDepthAllTime` shows meaningful
- * population at the ceiling. */
-export const CHAIN_DEPTH_MAX_BUCKET = 5;
-
-/** Day-grain window for the cohort funnel's "activated within 30
- * days" stage. Pinned as a constant so the value can't drift apart
- * from the SDL contract on `SysAdminCohortFunnelPoint.activatedD30`
- * via a bare magic number in the ms math. */
-export const COHORT_ACTIVATION_WINDOW_DAYS = 30;
-
-/**
- * Approximate days-per-month conversion used to translate the
- * operator-facing `minMonthsIn` (calendar months) into the internal
- * day-count check on `daysIn`. 30 is a deliberate approximation:
- * matches the boundaries used by `tenureDistribution`'s daysIn
- * buckets (lt1Month = daysIn < 30) so the stage classifier and the
- * tenure-distribution chart agree on what "1 month" means.
- */
-const DAYS_PER_MONTH_APPROX = 30;
-
-/**
- * Single source of truth for "what stage is this member in", used by
- * both `computeStageCounts` (cumulative semantics: tier1 ⊂ tier2) and
- * `computeStageBreakdown` (disjoint buckets). Centralising the rules
- * here prevents the two methods drifting when a new axis is added —
- * the `minMonthsIn` floor was almost added to one and not the other
- * during the issue #918 work.
- *
- * The tenure floor is checked against `daysIn`, not `monthsIn`,
- * because monthsIn is calendar-month-based: a member who joined on
- * Jan 31 and is observed on Feb 1 has `monthsIn = 2` (2 calendar
- * months touched) but only `daysIn = 2` (2 actual days). Using
- * monthsIn here would let that 2-day member sail past a
- * `minMonthsIn = 2` filter, defeating the artifact guard. daysIn
- * with a 30 d/month approximation matches the `tenureDistribution`
- * bucket semantics and treats both edges symmetrically.
- *
- * Classification:
- *   latent      — never donated (donationOutMonths === 0)
- *   habitual    — userSendRate >= tier1 AND daysIn >= minMonthsIn × 30
- *   regular     — userSendRate >= tier2 AND daysIn >= minMonthsIn × 30
- *                 (and not habitual, since we check tier1 first)
- *   occasional  — donated, but either below tier2 OR below the
- *                 tenure floor (the latter is the short-tenure
- *                 artifact guard: a brand-new member who donated
- *                 once cannot be elevated above occasional)
- */
-export type MemberClassification = "habitual" | "regular" | "occasional" | "latent";
-
-export function classifyMember(
-  m: SysAdminMemberStatsRow,
-  thresholds: SegmentThresholds,
-): MemberClassification {
-  if (m.donationOutMonths === 0) return "latent";
-  // tier1 / tier2 require BOTH the rate threshold AND the tenure
-  // floor. A donating-but-too-new member falls through to
-  // "occasional" — they've shown some activity, but we don't have
-  // enough tenure data to elevate their classification.
-  if (m.daysIn < thresholds.minMonthsIn * DAYS_PER_MONTH_APPROX) return "occasional";
-  if (m.userSendRate >= thresholds.tier1) return "habitual";
-  if (m.userSendRate >= thresholds.tier2) return "regular";
-  return "occasional";
-}
-
-/**
- * "Has this member donated before but gone quiet?" — the
- * latent-vs-dormant distinction operators care about for choosing
- * between onboarding and re-engagement interventions.
- *
- * False for never-donated members (latent / passive), regardless
- * of how long ago they joined. True only when the member's most
- * recent DONATION is strictly older than the threshold; equality
- * is treated as "still active" so a member who donated exactly
- * `dormantThresholdDays` days ago does NOT qualify.
- *
- * The cutoff is computed off the JST-truncated asOf so the
- * comparison stays deterministic regardless of the request's
- * wall-clock time. `lastDonationDay` is a JST calendar day at
- * UTC 00:00 (PostgreSQL `::date` cast); truncating asOf the same
- * way is the only way the strict-less-than has a stable boundary
- * — without it, a member who donated on the cutoff day would be
- * (incorrectly) dormant whenever `asOf` carried a nonzero
- * time-of-day component (i.e. essentially always in production).
- */
-export function isDormant(
-  m: SysAdminMemberStatsRow,
-  asOf: Date,
-  dormantThresholdDays: number,
-): boolean {
-  if (m.donationOutMonths === 0 || m.lastDonationDay === null) return false;
-  const cutoff = addDays(truncateToJstDate(asOf), -dormantThresholdDays);
-  return m.lastDonationDay < cutoff;
-}
-
-export type WeeklyRetentionPoint = {
-  weekStart: Date;
-  retainedSenders: number;
-  churnedSenders: number;
-  returnedSenders: number;
-  newMembers: number;
-  communityActivityRate: number | null;
-};
-
-export type MonthlyCohortPoint = {
-  cohortMonthStart: Date;
-  cohortSize: number;
-  retentionM1: number | null;
-  retentionM3: number | null;
-  retentionM6: number | null;
-};
-
-/**
- * One cohort's funnel progression. Backs
- * `SysAdminCommunityDetailPayload.cohortFunnel`. Computed by
- * `SysAdminService.computeCohortFunnel` over the already-fetched
- * `members` set; no separate repository call.
- */
-export type SysAdminCohortFunnelPoint = {
-  cohortMonth: Date;
-  acquired: number;
-  activatedD30: number;
-  repeated: number;
-  habitual: number;
-};
-
-export type SortField = "SEND_RATE" | "MONTHS_IN" | "DONATION_OUT_MONTHS" | "TOTAL_POINTS_OUT";
-export type SortOrder = "ASC" | "DESC";
-
-export type MemberListParams = {
-  minSendRate?: number | null;
-  maxSendRate?: number | null;
-  minMonthsIn?: number | null;
-  minDonationOutMonths?: number | null;
-  sortField: SortField;
-  sortOrder: SortOrder;
-  limit: number;
-  // Pre-decoded offset; the GraphQL → number wire-format step lives
-  // in `SysAdminConverter.parseMemberListCursor` so the service
-  // operates on internal form only.
-  cursor?: number;
-};
-
-export type MemberListResult = {
-  users: SysAdminMemberStatsRow[];
-  hasNextPage: boolean;
-  // Internal form. The GraphQL `nextCursor: String | null` is built
-  // by the presenter, which owns the internal → wire-format step.
-  nextOffset: number | null;
-};
+// Re-export moved symbols so existing importers via
+// `@/application/domain/sysadmin/service` continue to resolve.
+// Direct imports from classifiers/aggregations/pagination are
+// migrated in the next commit; the re-exports get pruned then.
+export {
+  classifyMember,
+  isDormant,
+  type MemberClassification,
+  type SegmentThresholds,
+  DEFAULT_SEGMENT_THRESHOLDS,
+  MIN_MIN_MONTHS_IN,
+  MAX_MIN_MONTHS_IN,
+} from "@/application/domain/sysadmin/classifiers";
+export {
+  type StageCounts,
+  type StageBucketStats,
+  type StageBreakdown,
+  type TenureHistogramBucket,
+  type TenureDistribution,
+  type SysAdminCohortFunnelPoint,
+  type WeeklyRetentionPoint,
+  type MonthlyCohortPoint,
+  TENURE_MONTHLY_BUCKETS,
+  CHAIN_DEPTH_MAX_BUCKET,
+  COHORT_ACTIVATION_WINDOW_DAYS,
+} from "@/application/domain/sysadmin/aggregations";
+export {
+  type SortField,
+  type SortOrder,
+  type MemberListParams,
+  type MemberListResult,
+  MAX_LIMIT,
+} from "@/application/domain/sysadmin/pagination";
 
 export type AlertFlags = {
   churnSpike: boolean;
@@ -334,18 +139,6 @@ export const DEFAULT_WINDOW_MONTHS = 10;
  * analytics need the spec references today.
  */
 export const MAX_WINDOW_MONTHS = 36;
-/**
- * Member-list page-size cap. Raised from the historical 200 to 1000
- * so client-side aggregations that span the full membership (e.g.
- * the L2 "受領→送付 転換率" / recipient-to-sender conversion rate
- * derived from `SysAdminMemberRow.uniqueDonationSenders` +
- * `totalPointsOut`) can pull a single page without N round-trips
- * for typical communities. Communities larger than 1000 members
- * still need cursor pagination — the cap exists to prevent a
- * malformed/hostile request from materialising the full member set
- * for an arbitrarily large community in one response.
- */
-export const MAX_LIMIT = 1000;
 export const ACTIVE_DROP_THRESHOLD = -0.2; // month-over-month fraction
 export const NO_NEW_MEMBERS_WINDOW_DAYS = 14;
 
@@ -416,17 +209,6 @@ async function runBatched<T, R>(
   return results;
 }
 
-/**
- * Fraction of the community that sent DONATION in a window:
- * `senderCount / totalMembers`, or 0 when the community had no members
- * during the window. Appears verbatim in several orchestrators
- * (getMonthActivityWithPrev, getAlerts, computeActivityRate3mAvg); one
- * helper keeps the "divide-by-zero ⇒ 0" convention in a single place.
- */
-function rateOf(senderCount: number, totalMembers: number): number {
-  return totalMembers === 0 ? 0 : senderCount / totalMembers;
-}
-
 @injectable()
 export default class SysAdminService {
   constructor(
@@ -440,263 +222,39 @@ export default class SysAdminService {
   ) {}
 
   // ==========================================================================
-  // Pure classification / aggregation helpers
-  //
-  // Split out from the orchestrator methods below so they're easy to
-  // unit-test without any database or context plumbing. Every one is a
-  // pure function of `members`, `thresholds`, and (for alerts) the
-  // pre-fetched rows.
+  // Pure classification / aggregation helpers — temporary delegations
+  // to the new module functions. Removed in the test-split commit;
+  // callers migrate to the module-level imports between then and now.
   // ==========================================================================
 
   computeStageCounts(
     members: SysAdminMemberStatsRow[],
     thresholds: SegmentThresholds,
   ): StageCounts {
-    let tier1Count = 0;
-    let tier2Count = 0;
-    let activeCount = 0;
-    let passiveCount = 0;
-    // Rules live in classifyMember so they stay in lockstep with
-    // computeStageBreakdown. The translation from disjoint
-    // classifications to the cumulative tier counts (habitual is also
-    // tier1 AND tier2; regular is also tier2) happens here.
-    for (const m of members) {
-      const c = classifyMember(m, thresholds);
-      if (c === "latent") {
-        passiveCount++;
-        continue;
-      }
-      activeCount++;
-      if (c === "habitual") {
-        tier1Count++;
-        tier2Count++;
-      } else if (c === "regular") {
-        tier2Count++;
-      }
-      // "occasional" contributes only to activeCount.
-    }
-    return { total: members.length, tier1Count, tier2Count, activeCount, passiveCount };
+    return computeStageCountsImpl(members, thresholds);
   }
 
-  /**
-   * Four-bucket stage breakdown. Unlike `computeStageCounts` these
-   * buckets are disjoint (`habitual`, `regular`, `occasional`, `latent`)
-   * so `pct` sums to 1.0. `pointsContributionPct` is the share of the
-   * community's total DONATION points-out attributed to each bucket.
-   */
   computeStageBreakdown(
     members: SysAdminMemberStatsRow[],
     thresholds: SegmentThresholds,
   ): StageBreakdown {
-    const totalMembers = members.length;
-    const totalPointsOut = members.reduce<bigint>((acc, m) => acc + m.totalPointsOut, BigInt(0));
-
-    const buckets: Record<MemberClassification, SysAdminMemberStatsRow[]> = {
-      habitual: [],
-      regular: [],
-      occasional: [],
-      latent: [],
-    };
-    // Bucket key === classification, so this collapses to a single
-    // dispatch per row. Disjoint semantics (a habitual member is NOT
-    // also in regular) come naturally from the classifyMember
-    // contract; the translation to cumulative counts lives in
-    // computeStageCounts.
-    for (const m of members) {
-      buckets[classifyMember(m, thresholds)].push(m);
-    }
-
-    const summarize = (rows: SysAdminMemberStatsRow[]): StageBucketStats => {
-      const count = rows.length;
-      if (count === 0) {
-        return {
-          count: 0,
-          pct: 0,
-          pointsContributionPct: 0,
-          avgSendRate: 0,
-          avgMonthsIn: 0,
-        };
-      }
-      // Single-pass reduce over rows so we visit the bucket once
-      // instead of three times. At realistic bucket sizes the
-      // difference is negligible, but it keeps the intent ("summarise
-      // this bucket") in one place.
-      const { sumSendRate, sumMonthsIn, sumPointsOut } = rows.reduce(
-        (acc, r) => ({
-          sumSendRate: acc.sumSendRate + r.userSendRate,
-          sumMonthsIn: acc.sumMonthsIn + r.monthsIn,
-          sumPointsOut: acc.sumPointsOut + r.totalPointsOut,
-        }),
-        { sumSendRate: 0, sumMonthsIn: 0, sumPointsOut: BigInt(0) },
-      );
-      // Route both sides of pointsContributionPct through
-      // bigintToSafeNumber so an extreme community's cumulative
-      // points total surfaces as a RangeError instead of silently
-      // truncating through `Number(bigint)`. The ratio itself is a
-      // [0, 1] fraction regardless of magnitude.
-      const pointsContributionPct =
-        totalPointsOut === BigInt(0)
-          ? 0
-          : bigintToSafeNumber(sumPointsOut) / bigintToSafeNumber(totalPointsOut);
-      return {
-        count,
-        pct: totalMembers === 0 ? 0 : count / totalMembers,
-        pointsContributionPct,
-        avgSendRate: sumSendRate / count,
-        avgMonthsIn: sumMonthsIn / count,
-      };
-    };
-
-    return {
-      habitual: summarize(buckets.habitual),
-      regular: summarize(buckets.regular),
-      occasional: summarize(buckets.occasional),
-      // The latent bucket's pointsContributionPct is always 0 by
-      // definition (latent === never-donated), so summarize() returns 0
-      // naturally without a special case.
-      latent: summarize(buckets.latent),
-    };
+    return computeStageBreakdownImpl(members, thresholds);
   }
 
-  /**
-   * Bucket members into 4 mutually exclusive tenure ranges by their
-   * `daysIn` (JST calendar-day tenure). Backs
-   * `SysAdminCommunityOverview.tenureDistribution` so the L1
-   * dashboard can render community age structure (new-heavy,
-   * established, etc.) without an L2 round-trip per community.
-   *
-   * Boundaries (left-inclusive, right-exclusive):
-   *   < 30 days       → lt1Month
-   *   [30, 90)        → m1to3Months
-   *   [90, 365)       → m3to12Months
-   *   >= 365          → gte12Months
-   *
-   * Day-based (not month-based) so a brand-new member never gets
-   * lifted into the "1 month" bucket purely because of `monthsIn`'s
-   * GREATEST(1, ...) floor.
-   */
   computeTenureDistribution(members: SysAdminMemberStatsRow[]): TenureDistribution {
-    let lt1Month = 0;
-    let m1to3Months = 0;
-    let m3to12Months = 0;
-    let gte12Months = 0;
-    // monthlyHistogram pre-allocates all 13 buckets (0..12) so the
-    // returned array always has a stable shape regardless of input.
-    // The 12 bucket aggregates daysIn >= 365 (= 12 calendar months
-    // by the same threshold the coarse `gte12Months` uses) so the
-    // two stay consistent: any member counted in `gte12Months` is
-    // also in histogram[12], and vice versa. Buckets 0..11 cover
-    // the [0, 365) range — bucket k = floor(daysIn / 30), clamped
-    // to 11 so 360..364 days do NOT silently land in the 12+
-    // aggregator (otherwise they'd be in histogram[12] but coarse
-    // `m3to12Months`, an asymmetry that would confuse downstream
-    // consumers stacking the histogram against the coarse bars).
-    const monthlyHistogram: TenureHistogramBucket[] = Array.from(
-      { length: TENURE_MONTHLY_BUCKETS },
-      (_, monthsIn) => ({ monthsIn, count: 0 }),
-    );
-    for (const m of members) {
-      if (m.daysIn < 30) lt1Month++;
-      else if (m.daysIn < 90) m1to3Months++;
-      else if (m.daysIn < 365) m3to12Months++;
-      else gte12Months++;
-      // Members with daysIn < 0 (data anomaly — shouldn't occur
-      // because findMemberStats clamps daysIn to >= 1) are
-      // already counted in lt1Month above (daysIn < 0 < 30), so
-      // they also flow into histogram bucket 0 via the
-      // Math.max(0, ...) clamp below. Without that clamp,
-      // floor(negative / 30) would land in a negative array
-      // index and silently corrupt the histogram. Counting both
-      // representations keeps the documented derivation
-      // invariant (lt1Month == histogram[0]) intact.
-      const bucket =
-        m.daysIn >= 365
-          ? TENURE_MONTHLY_BUCKETS - 1
-          : Math.min(Math.max(Math.floor(m.daysIn / 30), 0), TENURE_MONTHLY_BUCKETS - 2);
-      monthlyHistogram[bucket].count++;
-    }
-    return { lt1Month, m1to3Months, m3to12Months, gte12Months, monthlyHistogram };
+    return computeTenureDistributionImpl(members);
   }
 
-  /**
-   * Count of members who donated at some point in the past but
-   * whose most recent DONATION is older than `dormantThresholdDays`.
-   * Excludes never-donated members (they're "latent" / passiveCount,
-   * not dormant — different intervention surface).
-   *
-   * Pure function over the member rows already fetched for the
-   * dashboard; no extra SQL.
-   */
   computeDormantCount(
     members: SysAdminMemberStatsRow[],
     asOf: Date,
     dormantThresholdDays: number,
   ): number {
-    let dormant = 0;
-    for (const m of members) {
-      if (isDormant(m, asOf, dormantThresholdDays)) dormant++;
-    }
-    return dormant;
+    return computeDormantCountImpl(members, asOf, dormantThresholdDays);
   }
 
-  /**
-   * Filter + sort + page a pre-fetched member list. Pagination uses an
-   * offset-encoded cursor (`"n"`) so the caller can resume from row N
-   * without re-running the underlying aggregation. At <6 communities
-   * with a few hundred members each, in-memory paging is cheaper than
-   * a new SQL round-trip per page.
-   */
   paginateMembers(members: SysAdminMemberStatsRow[], params: MemberListParams): MemberListResult {
-    const filtered = members.filter((m) => {
-      if (params.minSendRate != null && m.userSendRate < params.minSendRate) return false;
-      if (params.maxSendRate != null && m.userSendRate > params.maxSendRate) return false;
-      if (params.minMonthsIn != null && m.monthsIn < params.minMonthsIn) return false;
-      if (params.minDonationOutMonths != null && m.donationOutMonths < params.minDonationOutMonths)
-        return false;
-      return true;
-    });
-
-    const sign = params.sortOrder === "ASC" ? 1 : -1;
-    const numericKey = (m: SysAdminMemberStatsRow): number => {
-      switch (params.sortField) {
-        case "SEND_RATE":
-          return m.userSendRate;
-        case "MONTHS_IN":
-          return m.monthsIn;
-        case "DONATION_OUT_MONTHS":
-          return m.donationOutMonths;
-        case "TOTAL_POINTS_OUT":
-          // Handled in the bigint branch below; placeholder here so
-          // the switch is exhaustive.
-          return 0;
-      }
-    };
-    const sorted = [...filtered].sort((a, b) => {
-      let cmp: number;
-      if (params.sortField === "TOTAL_POINTS_OUT") {
-        // Compare bigints directly — Number(bigint) would silently
-        // lose precision past Number.MAX_SAFE_INTEGER and quietly
-        // mis-order the leaderboard for extreme donors.
-        if (a.totalPointsOut < b.totalPointsOut) cmp = -1;
-        else if (a.totalPointsOut > b.totalPointsOut) cmp = 1;
-        else cmp = 0;
-      } else {
-        cmp = numericKey(a) - numericKey(b);
-      }
-      if (cmp !== 0) return sign * cmp;
-      // Stable secondary sort on userId to make the cursor deterministic.
-      return a.userId.localeCompare(b.userId);
-    });
-
-    const start = params.cursor ?? 0;
-    const limit = Math.min(Math.max(params.limit, 1), MAX_LIMIT);
-    const page = sorted.slice(start, start + limit);
-    const hasNextPage = start + limit < sorted.length;
-    return {
-      users: page,
-      hasNextPage,
-      nextOffset: hasNextPage ? start + limit : null,
-    };
+    return paginateMembersImpl(members, params);
   }
 
   // ==========================================================================
@@ -734,81 +292,13 @@ export default class SysAdminService {
     );
   }
 
-  /**
-   * Per-cohort send-funnel progression for the L3 deep-dive. Pure
-   * function over `members` (already-fetched in the L2 usecase) +
-   * `asOf` + `windowMonths` + `thresholds` — no extra DB scan.
-   *
-   * Cohort = JST month start of the member's `joinedAt`. Members
-   * outside the trailing windowMonths range are dropped. Stage
-   * classification:
-   *
-   *   acquired      — every cohort member counts
-   *   activatedD30  — has firstDonationDay AND
-   *                   firstDonationDay - joinedAt < 30 days
-   *   repeated      — donationOutMonths >= 2
-   *   habitual      — classifyMember(...) === "habitual"
-   *
-   * activatedD30 / repeated / habitual are JOINED-at-asOf scoped
-   * because `members` is itself JOINED-at-asOf (findMemberStats
-   * applies the membership filter).
-   */
   computeCohortFunnel(
     members: SysAdminMemberStatsRow[],
     asOf: Date,
     windowMonths: number,
     thresholds: SegmentThresholds,
   ): SysAdminCohortFunnelPoint[] {
-    // Build the cohort-month axis: [asOf month - (windowMonths-1), ..., asOf month].
-    // Same orientation as `monthlyActivityTrend` (newest last) so the
-    // client can render a single x-axis across both series.
-    const cohortKeys: Date[] = [];
-    for (let i = windowMonths - 1; i >= 0; i--) {
-      cohortKeys.push(jstMonthStartOffset(asOf, -i));
-    }
-    const buckets = new Map<number, SysAdminCohortFunnelPoint>();
-    for (const m of cohortKeys) {
-      buckets.set(m.getTime(), {
-        cohortMonth: m,
-        acquired: 0,
-        activatedD30: 0,
-        repeated: 0,
-        habitual: 0,
-      });
-    }
-    const earliest = cohortKeys[0]?.getTime() ?? 0;
-    const latest = cohortKeys[cohortKeys.length - 1]?.getTime() ?? 0;
-    const activationCutoffMs = COHORT_ACTIVATION_WINDOW_DAYS * 24 * 60 * 60 * 1000;
-    for (const member of members) {
-      const cohort = jstMonthStart(member.joinedAt).getTime();
-      if (cohort < earliest || cohort > latest) continue;
-      const bucket = buckets.get(cohort);
-      if (!bucket) continue;
-      bucket.acquired++;
-      // Both sides are JST-day grain: firstDonationDay is already
-      // a JST date encoded at UTC midnight (see findMemberStats),
-      // and joinedAt is truncated to its JST date here so the
-      // "30 days within join" comparison is symmetric. Without
-      // truncation the joinedAt time-of-day component skewed the
-      // comparison by up to ±9h (JST offset), giving false
-      // positives / negatives at the 30-day boundary. The day-
-      // grain comparison matches the SDL "within 30 days of join"
-      // wording the operator UI exposes.
-      if (
-        member.firstDonationDay !== null &&
-        member.firstDonationDay.getTime() - truncateToJstDate(member.joinedAt).getTime() <
-          activationCutoffMs
-      ) {
-        bucket.activatedD30++;
-      }
-      if (member.donationOutMonths >= 2) {
-        bucket.repeated++;
-      }
-      if (classifyMember(member, thresholds) === "habitual") {
-        bucket.habitual++;
-      }
-    }
-    return cohortKeys.map((k) => buckets.get(k.getTime())!);
+    return computeCohortFunnelImpl(members, asOf, windowMonths, thresholds);
   }
 
   async getMonthlyActivity(
@@ -1058,16 +548,8 @@ export default class SysAdminService {
     };
   }
 
-  /** 3-month trailing average of communityActivityRate, ending at
-   * asOf's month (inclusive). null when <3 months of trend data exist. */
   computeActivityRate3mAvg(trend: SysAdminMonthlyActivityRow[]): number | null {
-    if (trend.length < 3) return null;
-    const last3 = trend.slice(-3);
-    const sumRates = last3.reduce(
-      (acc, r) => acc + rateOf(r.senderCount, r.totalMembersEndOfMonth),
-      0,
-    );
-    return sumRates / 3;
+    return computeActivityRate3mAvgImpl(trend);
   }
 
   /**
